# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface CategoryStat {
  category: string;
  orders: number;
  revenue: number;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}

export interface AnalyticsSummary {
  totalRevenue: number;
  totalOrders: number;
  activeOrders: number;
  outOfStockItems: number;
  categoryBreakdown: CategoryStat[];
  dailyRevenue: DailyRevenue[];
}

export interface AdminMenuItem {
  id: string;
  name: string;
  description: string;
  price: number;
  category: string;
  available: boolean;
  spicy?: boolean;
  popular?: boolean;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  offerPercentage?: number | null;
  /** @nullable */
  offerLabel?: string | null;
  offerActive?: boolean;
  /** @nullable */
  offerStartDate?: string | null;
  /** @nullable */
  offerEndDate?: string | null;
}

export interface AdminMenuItemInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
  spicy?: boolean;
  popular?: boolean;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  offerPercentage?: number | null;
  /** @nullable */
  offerLabel?: string | null;
  offerActive?: boolean;
  /** @nullable */
  offerStartDate?: string | null;
  /** @nullable */
  offerEndDate?: string | null;
}

export interface AdminMenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  available?: boolean;
  spicy?: boolean;
  popular?: boolean;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  offerPercentage?: number | null;
  /** @nullable */
  offerLabel?: string | null;
  offerActive?: boolean;
  /** @nullable */
  offerStartDate?: string | null;
  /** @nullable */
  offerEndDate?: string | null;
}

export interface AdminOrderItem {
  name: string;
  quantity: number;
  price: number;
}

export interface AdminOrder {
  id: string;
  customerName: string;
  customerPhone: string;
  branch: string;
  orderType: string;
  items: AdminOrderItem[];
  total: number;
  status: string;
  createdAt: string;
}

export interface AdminCategory {
  id: string;
  name: string;
  slug: string;
  itemCount: number;
}

export interface AdminCategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface OrderStatusUpdate {
  status: string;
}