# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/menu:
    get:
      operationId: listAdminMenuItems
      tags: [admin-menu]
      summary: List all menu items
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminMenuItem"
    post:
      operationId: createAdminMenuItem
      tags: [admin-menu]
      summary: Add a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMenuItemInput"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMenuItem"
  /admin/menu/{id}:
    patch:
      operationId: updateAdminMenuItem
      tags: [admin-menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMenuItemUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMenuItem"
        "404":
          description: Not found
    delete:
      operationId: deleteAdminMenuItem
      tags: [admin-menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [admin-menu]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCategory"
    post:
      operationId: createAdminCategory
      tags: [admin-menu]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategory"
        "409":
          description: Category already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/categories/{id}:
    delete:
      operationId: deleteAdminCategory
      tags: [admin-menu]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Category has items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/orders/{id}/status:
    patch:
      operationId: updateAdminOrderStatus
      tags: [admin-orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"
        "404":
          description: Not found
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AdminMenuItem:
      type: object
      required: [id, name, description, price, category, available]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        available:
          type: boolean
        spicy:
          type: boolean
        popular:
          type: boolean
        calories:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        offerPercentage:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
        offerLabel:
          type: ["string", "null"]
        offerActive:
          type: boolean
        offerStartDate:
          type: ["string", "null"]
        offerEndDate:
          type: ["string", "null"]
    AdminMenuItemInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
        spicy:
          type: boolean
        popular:
          type: boolean
        calories:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        offerPercentage:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
        offerLabel:
          type: ["string", "null"]
        offerActive:
          type: boolean
        offerStartDate:
          type: ["string", "null"]
        offerEndDate:
          type: ["string", "null"]
    AdminMenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        available:
          type: boolean
        spicy:
          type: boolean
        popular:
          type: boolean
        calories:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        offerPercentage:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
        offerLabel:
          type: ["string", "null"]
        offerActive:
          type: boolean
        offerStartDate:
          type: ["string", "null"]
        offerEndDate:
          type: ["string", "null"]
    AdminOrder:
      type: object
      required: [id, customerName, customerPhone, branch, orderType, items, total, status, createdAt]
      properties:
        id:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        branch:
          type: string
        orderType:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderItem"
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
    AdminOrderItem:
      type: object
      required: [name, quantity, price]
      properties:
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
    AdminCategory:
      type: object
      required: [id, name, slug, itemCount]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        itemCount:
          type: integer
    AdminCategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string