# lib/api-zod/src/generated/types/adminMenuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminMenuItem {
  id: string;
  name: string;
  description: string;
  price: number;
  category: string;
  available: boolean;
  spicy?: boolean;
  popular?: boolean;
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  offerPercentage?: number | null;
  /** @nullable */
  offerLabel?: string | null;
  offerActive?: boolean;
  /** @nullable */
  offerStartDate?: string | null;
  /** @nullable */
  offerEndDate?: string | null;
}